Erik + Battlefield;  Ben +Corneria
Battlefield done, but only Colliders, used mesh Colliders when Box
Colliders would have worked and been cheaper.
Corneria, Colliders and basic script for firing 'LAZORS' out of the ship
guns.  Needs to loop and the ship needs to 'fly' through the stage.
(recommend leaving the ship in place, and moving the stage itself.)